# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Authorization and Сontact-book API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api"
    }
  ],
  "paths": {
...
    "/contacts/{contact_Id}": {
      "get": {
        "summary": "Get contact by Id",
        "tags": ["Сontacts"],
        "operationId": "getContactById",
        "parameters": [
          {
            "name": "contact_Id",
            "in": "path",
            "required": true,
            "description": "id contact",
            "schema": {
              "type": "string"
            },
            "example": "651db689e2e663b6128bf7b3"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "contact": {
                      "$ref": "#/components/schemas/Contact"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Sorry, is not valid id."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "summary": "Update contact by Id",
        "tags": ["Сontacts"],
        "operationId": "putContactById",
        "parameters": [
          {
            "name": "contact_Id",
            "in": "path",
            "required": true,
            "description": "id contact",
            "schema": {
              "type": "string"
            },
            "example": "651db689e2e663b6128bf7b3"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Bob Jonson"
                  },
                  "email": {
                    "type": "string",
                    "example": "gutum@sci.net"
                  },
                  "phone": {
                    "type": "string",
                    "example": "(666) 555-6666"
                  },
                  "avatar": {
                    "type": "string",
                    "example": "www.gravatar.coceae3375ff085a80d80e31e960da34",
                    "format": "binary",
                    "description": "The avatar image file to upload."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "updateContact": {
                      "$ref": "#/components/schemas/Contact"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Delete contact by Id",
        "tags": ["Сontacts"],
        "operationId": "deleteContactById",
        "parameters": [
          {
            "name": "contact_Id",
            "in": "path",
            "required": true,
            "description": "id contact",
            "schema": {
              "type": "string"
            },
            "example": "651db689e2e663b6128bf7b3"
          }
        ],
        "responses": {
          "200": {
            "description": "Succes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Delete success."
                    },
                    "deletedContact": {
                      "$ref": "#/components/schemas/Contact"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BadRequest",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Sorry, is not valid id."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Contact": {
        "type": "object",
        "required": ["name", "email", "phone"],
        "properties": {
          "_id": {
            "type": "string",
            "example": "651db689e2e663b6128bf7b3"
          },
          "name": {
            "type": "string",
            "example": "John Robson"
          },
          "email": {
            "type": "string",
            "example": "email@sci.net"
          },
          "phone": {
            "type": "string",
            "example": "(666) 555-6666"
          },
          "favorite": {
            "type": "boolean",
            "example": false
          },
          "cover": {
            "example": null
          }
        }
      },
...
      "NotFound": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Sorry, contact not found"
          }
        }
      },
      "BadRequest": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Validation error's description"
          }
        }
      },
      "Unauthorized": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Not authorized"
          }
        }
      },
      "ServerError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Server error"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}